write new Python opuscraft program
This new program synthesizes an Opus file according to a script. The
script pulls individual sounds from input files, groups them into words,
and writes them in the order specified by an output file. This appears
to be a very successful approach.

In testing we fit 14.86 s of audio in 1398 bytes, in a *.opus.gz file.
It contained several English phrases and was... relatively intelligible?
The phrases probably won't be understood without the aid of subtitles.